New HistoryCallback + distribution over multiple files
