+++ klax/_callbacks.py
@@ -16,7 +16,7 @@ import datetime
 import importlib
 import pickle
 import time
-from abc import ABC
+from abc import ABC, abstractmethod
 from collections.abc import Callable
 from pathlib import Path
 from typing import Any, Self
@@ -24,132 +24,7 @@ from typing import Any, Self
 import jax
 from jaxtyping import PyTree, PyTreeDef, Scalar
 
-
-class CallbackArgs:
-    """A callback argument designed to work in conjunction with [`klax.fit`][].
-
-    This class should not be instantiated directly. An instance of this class
-    is passed to every callback object in the fit function. When writing a
-    custom callback, use the properties of this class to access the current
-    model, optimizer state, training data, and validation data during training.
-
-    This class implements cached and lazy-evaluated values via property
-    methods. This means that properties like ``loss`` are only calculated if
-    they are used and are stored such that they are not calculated multiple
-    times.
-    """
-
-    step: int  #: Current step-count of the training.
-    time_on_last_update: float  #: Global time of the last :meth:`update` call.
-    data: PyTree  #: PyTree of the training data.
-    val_data: PyTree | None  #: PyTree of the validation data.
-    _treedef_model: PyTreeDef
-    _flat_model: list
-    _treedef_opt_state: PyTreeDef
-    _flat_opt_state: list
-    _cache: dict = {}
-    _get_loss: Callable[[PyTree, PyTree], Scalar]
-    _start_time: float
-
-    def __init__(
-        self,
-        get_loss: Callable[[PyTree, PyTree], Scalar],
-        treedef_model: PyTreeDef,
-        treedef_opt_state: PyTreeDef,
-        data: PyTree,
-        val_data: PyTree | None = None,
-    ):
-        """Initialize the callback arguments object.
-
-        Args:
-        get_loss: Function that takes a model and a batch of data and
-            returns the loss.
-        treedef_model: Tree structure of the model.
-        treedef_opt_state: Tree structure of the :py:mod:`optax` optimizer.
-        data: PyTree of the training data.
-        val_data: PyTree of the validation data. If None, no validation
-            loss is calculated and the property :py:attr:`val_loss` will
-            return None.
-
-        """
-        self.data = data
-        self.val_data = val_data
-        self._get_loss = get_loss
-        self._treedef_model = treedef_model
-        self._treedef_opt_state = treedef_opt_state
-
-    def update(self, flat_model: PyTree, flat_opt_state: PyTree, step: int):
-        """Update the object with the current model and optimizer state.
-
-        This method is called repeatedly in [`klax.fit`][].
-
-        Args:
-            flat_model: Flattened PyTree of the model.
-            flat_opt_state: Flattened PyTree of the `optax`
-                optimizer.
-            step: Current step-count of the training.
-
-        """
-        self._flat_model = flat_model
-        self._flat_opt_state = flat_opt_state
-        self.step = step
-        self.time_on_last_update = time.time()
-
-        # Clear cache
-        self._cache = {}
-
-    @staticmethod
-    def _lazy_evaluated_and_cached(fun: Callable[[Any], Any]) -> property:
-        """Turn a public method into a property.
-
-        The return value of ``fun`` is stored in the ``_cache`` dictionary of
-        the current object using the function name as key. If the name is
-        already in ``_cache`` then the cached value is simply returned,
-        without evaluating ``fun``.
-
-        Args:
-            fun: Method to wrap.
-
-        Returns:
-            Wrapped method as a property.
-
-        """
-        attr_name = fun.__name__
-
-        def wrapper(self: Self):
-            if attr_name not in self._cache:
-                self._cache.setdefault(attr_name, fun(self))
-            return self._cache.get(attr_name)
-
-        wrapper.__doc__ = fun.__doc__
-
-        return property(wrapper)
-
-    @_lazy_evaluated_and_cached
-    def model(self):
-        """Lazy-evaluated and cached model."""
-        return jax.tree_util.tree_unflatten(
-            self._treedef_model, self._flat_model
-        )
-
-    @_lazy_evaluated_and_cached
-    def opt_state(self):
-        """Lazy-evaluated and cached optimizer state."""
-        return jax.tree_util.tree_unflatten(
-            self._treedef_opt_state, self._flat_opt_state
-        )
-
-    @_lazy_evaluated_and_cached
-    def loss(self):
-        """Lazy-evaluated and cached training loss."""
-        return self._get_loss(self.model, self.data)
-
-    @_lazy_evaluated_and_cached
-    def val_loss(self) -> Scalar | None:
-        """Lazy-evaluated and cached validation loss."""
-        if self.val_data is None:
-            return None
-        return self._get_loss(self.model, self.val_data)
+from ._context import TrainingContext
 
 
 class Callback(ABC):
@@ -158,17 +33,166 @@ class Callback(ABC):
     Inherit from this class to create a custom callback.
     """
 
-    def __call__(self, cbargs: CallbackArgs) -> bool | None:
+    @abstractmethod
+    def __call__(self, ctx: TrainingContext) -> bool | None:
         """Call after each step during training."""
         pass
 
-    def on_training_end(self, cbargs: CallbackArgs) -> None:
+    @abstractmethod
+    def on_training_end(self, ctx: TrainingContext) -> None:
         """Call when training ends."""
         pass
 
-    def on_training_start(self, cbargs: CallbackArgs) -> None:
+    @abstractmethod
+    def on_training_start(self, ctx: TrainingContext) -> None:
         """Call when training starts."""
         pass
+
+
+# class CallbackArgs:
+#     """A callback argument designed to work in conjunction with [`klax.fit`][].
+
+#     This class should not be instantiated directly. An instance of this class
+#     is passed to every callback object in the fit function. When writing a
+#     custom callback, use the properties of this class to access the current
+#     model, optimizer state, training data, and validation data during training.
+
+#     This class implements cached and lazy-evaluated values via property
+#     methods. This means that properties like ``loss`` are only calculated if
+#     they are used and are stored such that they are not calculated multiple
+#     times.
+#     """
+
+#     step: int  #: Current step-count of the training.
+#     time_on_last_update: float  #: Global time of the last :meth:`update` call.
+#     data: PyTree  #: PyTree of the training data.
+#     val_data: PyTree | None  #: PyTree of the validation data.
+#     _treedef_model: PyTreeDef
+#     _flat_model: list
+#     _treedef_opt_state: PyTreeDef
+#     _flat_opt_state: list
+#     _cache: dict = {}
+#     _get_loss: Callable[[PyTree, PyTree], Scalar]
+#     _start_time: float
+
+#     def __init__(
+#         self,
+#         get_loss: Callable[[PyTree, PyTree], Scalar],
+#         treedef_model: PyTreeDef,
+#         treedef_opt_state: PyTreeDef,
+#         data: PyTree,
+#         val_data: PyTree | None = None,
+#     ):
+#         """Initialize the callback arguments object.
+
+#         Args:
+#         get_loss: Function that takes a model and a batch of data and
+#             returns the loss.
+#         treedef_model: Tree structure of the model.
+#         treedef_opt_state: Tree structure of the :py:mod:`optax` optimizer.
+#         data: PyTree of the training data.
+#         val_data: PyTree of the validation data. If None, no validation
+#             loss is calculated and the property :py:attr:`val_loss` will
+#             return None.
+
+#         """
+#         self.data = data
+#         self.val_data = val_data
+#         self._get_loss = get_loss
+#         self._treedef_model = treedef_model
+#         self._treedef_opt_state = treedef_opt_state
+
+#     def update(self, flat_model: PyTree, flat_opt_state: PyTree, step: int):
+#         """Update the object with the current model and optimizer state.
+
+#         This method is called repeatedly in [`klax.fit`][].
+
+#         Args:
+#             flat_model: Flattened PyTree of the model.
+#             flat_opt_state: Flattened PyTree of the `optax`
+#                 optimizer.
+#             step: Current step-count of the training.
+
+#         """
+#         self._flat_model = flat_model
+#         self._flat_opt_state = flat_opt_state
+#         self.step = step
+#         self.time_on_last_update = time.time()
+
+#         # Clear cache
+#         self._cache = {}
+
+#     @staticmethod
+#     def _lazy_evaluated_and_cached(fun: Callable[[Any], Any]) -> property:
+#         """Turn a public method into a property.
+
+#         The return value of ``fun`` is stored in the ``_cache`` dictionary of
+#         the current object using the function name as key. If the name is
+#         already in ``_cache`` then the cached value is simply returned,
+#         without evaluating ``fun``.
+
+#         Args:
+#             fun: Method to wrap.
+
+#         Returns:
+#             Wrapped method as a property.
+
+#         """
+#         attr_name = fun.__name__
+
+#         def wrapper(self: Self):
+#             if attr_name not in self._cache:
+#                 self._cache.setdefault(attr_name, fun(self))
+#             return self._cache.get(attr_name)
+
+#         wrapper.__doc__ = fun.__doc__
+
+#         return property(wrapper)
+
+#     @_lazy_evaluated_and_cached
+#     def model(self):
+#         """Lazy-evaluated and cached model."""
+#         return jax.tree_util.tree_unflatten(
+#             self._treedef_model, self._flat_model
+#         )
+
+#     @_lazy_evaluated_and_cached
+#     def opt_state(self):
+#         """Lazy-evaluated and cached optimizer state."""
+#         return jax.tree_util.tree_unflatten(
+#             self._treedef_opt_state, self._flat_opt_state
+#         )
+
+#     @_lazy_evaluated_and_cached
+#     def loss(self):
+#         """Lazy-evaluated and cached training loss."""
+#         return self._get_loss(self.model, self.data)
+
+#     @_lazy_evaluated_and_cached
+#     def val_loss(self) -> Scalar | None:
+#         """Lazy-evaluated and cached validation loss."""
+#         if self.val_data is None:
+#             return None
+#         return self._get_loss(self.model, self.val_data)
+
+
+# class Callback(ABC):
+#     """An abstract callback.
+
+#     Inherit from this class to create a custom callback.
+#     """
+
+#     def __call__(self, cbargs: CallbackArgs) -> bool | None:
+#         """Call after each step during training."""
+#         pass
+
+#     def on_training_end(self, cbargs: CallbackArgs) -> None:
+#         """Call when training ends."""
+#         pass
+
+#     def on_training_start(self, cbargs: CallbackArgs) -> None:
+#         """Call when training starts."""
+#         pass
 
 
 class HistoryCallback(Callback):
@@ -181,8 +205,6 @@ class HistoryCallback(Callback):
     steps: list  #: List of steps at which the losses were recorded.
     loss: list
     val_loss: list
-    last_start_time: float  # start time of the last training
-    last_end_time: float  # End time of the last training
     training_time: float = 0  # Total training time of all trainings
     verbose: bool
     step_offset: int = 0  # Potential offset due to previous trainings
@@ -211,44 +233,45 @@ class HistoryCallback(Callback):
             f"verbose={self.verbose})"
         )
 
-    def __call__(self, cbargs: CallbackArgs):
+    def __call__(self, ctx: TrainingContext):
         """Record the losses and step count.
 
         Called at each step during training.
         """
-        if cbargs.step % self.log_every == 0:
-            self.steps.append(self.step_offset + cbargs.step)
-            self.loss.append(cbargs.loss)
-            self.val_loss.append(cbargs.val_loss)
+        if ctx.step % self.log_every == 0:
+            self.steps.append(self.step_offset + ctx.step)
+            self.loss.append(ctx.loss)
+            self.val_loss.append(ctx.val_loss)
 
             # Print message
             if self.verbose:
-                message = f"Step: {cbargs.step}, Loss: {cbargs.loss:.3e}"
-                if cbargs.val_data is not None:
-                    message += f", Validation loss: {cbargs.val_loss:.3e}"
+                message = f"Step: {ctx.step}, Loss: {ctx.loss:.3e}"
+                if ctx.val_data is not None:
+                    message += f", Validation loss: {ctx.val_loss:.3e}"
                 print(message)
 
-    def on_training_start(self, cbargs: CallbackArgs):
+    def on_training_start(self, ctx: TrainingContext):
         """Initialize the training start time.
 
         Called at beginning of training.
         """
-        self.last_start_time = cbargs.time_on_last_update
+        assert ctx.start_time is not None
+        self.last_start_time = ctx.start_time
         if self.steps:
             # If there are already steps, we assume that this is a continuation
             # of a training.
             self.step_offset = self.steps[-1]
         else:
-            self(cbargs)
+            self(ctx)
 
-    def on_training_end(self, cbargs: CallbackArgs):
+    def on_training_end(self, ctx: TrainingContext):
         """Record the training end time and the last optimizer state.
 
         Called at end of training.
         """
-        self.last_end_time = cbargs.time_on_last_update
-        self.training_time += self.last_end_time - self.last_start_time
-        self.last_opt_state = cbargs.opt_state
+        assert ctx.total_time is not None
+        self.training_time += ctx.total_time
+        self.last_opt_state = ctx.opt_state
         if self.verbose:
             print(
                 f"Training took: {
