+++ klax/_losses.py
@@ -14,109 +14,65 @@
 
 import typing
 from abc import abstractmethod
-from collections.abc import Sequence
 from typing import Any, Protocol
 
+import equinox as eqx
 import jax
 import jax.numpy as jnp
 from jaxtyping import PyTree, Scalar
 
 
 @typing.runtime_checkable
-class Loss(Protocol):
-    """An abstract callable loss object.
-
-    It can be used to build custom losses that can be passed to [`klax.fit`][].
-
-    Example:
-        A simple custom loss that computes the mean squared error between
-        the predicted values `y_pred` and true values `y` for in inputs `x` may
-        be implemented as follows:
-
-        ```python
-        >>> def mse(model, data, batch_axis=0):
-        ...    x, y = data
-        ...    if isinstance(batch_axis, tuple):
-        ...        in_axes = batch_axis[0]
-        ...    else:
-        ...        in_axes = batch_axis
-        ...    y_pred = jax.vmap(model, in_axes=(in_axes,))(x)
-        ...    return jnp.mean(jnp.square(y_pred - y))
-        ```
-
-        Note that, since we a aim to provide a maximum of flexibility the users
-        have to take care of applying `jax.vmap` to the model themselves.
-
-    """
-
+class ValueFn(Protocol):
     @abstractmethod
-    def __call__(
-        self,
-        model: PyTree,
-        data: PyTree,
-        batch_axis: int | None | Sequence[Any],
-    ) -> Scalar:
-        """Abstract method to compute the loss for a given model and data.
-
-        Args:
-            model: The model parameters or structure to evaluate the loss.
-            data: The input data or structure used for loss computation.
-            batch_axis: Specifies the axis or axes corresponding to the batch
-                dimension in the data. Can be an integer, None, or a sequence
-                of values.
-
-        Returns:
-            Scalar: The computed loss value.
-
-        """
-        ...
+    def __call__(self, model: PyTree, data: PyTree) -> Scalar:
+        pass
 
 
-class MSE(Loss):
-    """Mean squared error for a tuple of data `(x, y)`.
+@typing.runtime_checkable
+class ValueAndGradFn(Protocol):
+    @abstractmethod
+    def __call__(self, model: PyTree, data: PyTree) -> tuple[Scalar, PyTree]:
+        pass
 
-    The inputs `x` and the outputs `y` are expected to have the same batch axis
-    and equal length along that axis.
-    """
 
-    def __call__(
-        self,
-        model: PyTree,
-        data: PyTree,
-        batch_axis: int | None | Sequence[Any] = 0,
-    ) -> Scalar:
+class LossFactory(Protocol):
+    @abstractmethod
+    def __call__(self, batch_axis) -> tuple[ValueFn, ValueAndGradFn]:
+        pass
+
+
+def mse(batch_axis) -> tuple[ValueFn, ValueAndGradFn]:
+    def value_fn(model: PyTree, data: PyTree) -> Scalar:
         x, y = data
         if isinstance(batch_axis, tuple):
             in_axes = batch_axis[0]
         else:
             in_axes = batch_axis
-        y_pred = jax.vmap(model, in_axes=(in_axes,))(x)
+        y_pred = eqx.filter_vmap(model, in_axes=(in_axes,))(x)
         return jnp.mean(jnp.square(y_pred - y))
 
+    def value_and_grad_fn(
+        model: PyTree, data: PyTree
+    ) -> tuple[Scalar, PyTree]:
+        return eqx.filter_value_and_grad(value_fn)(model, data)
 
-mse = MSE()
+    return value_fn, value_and_grad_fn
 
 
-class MAE(Loss):
-    """Mean absolute error for a tuple of data `(x, y)`.
-
-    The inputs `x` and the outputs `y` are expected to have the same batch axis
-    and equal length along that axis.
-    """
-
-    def __call__(
-        self,
-        model: PyTree,
-        data: PyTree,
-        batch_axis: int | None | Sequence[Any] = 0,
-    ) -> Scalar:
+def mae(batch_axis) -> tuple[ValueFn, ValueAndGradFn]:
+    def value_fn(model: PyTree, data: PyTree) -> Scalar:
         x, y = data
         if isinstance(batch_axis, tuple):
             in_axes = batch_axis[0]
         else:
             in_axes = batch_axis
-        y_pred = jax.vmap(model, in_axes=(in_axes,))(x)
+        y_pred = eqx.filter_vmap(model, in_axes=(in_axes,))(x)
         return jnp.mean(jnp.abs(y_pred - y))
 
+    def value_and_grad_fn(
+        model: PyTree, data: PyTree
+    ) -> tuple[Scalar, PyTree]:
+        return eqx.filter_value_and_grad(value_fn)(model, data)
 
-mae = MAE()
+    return value_fn, value_and_grad_fn
